football-data-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "football-data-pipeline"
version = "0.1.0"
description = "End-to-end data engineering pipeline for football statistics from understat.com"
authors = ["Habeeb Babatunde <babatundehabeeb2@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/habeeb3579/football-data-pipeline"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
# Data extraction
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
selenium = "^4.8.3"
lxml = "^4.9.2"
html5lib = "^1.1"

# Data processing & transformation
pandas = "^2.0.0"
numpy = "^1.24.2"
dlt = "^0.3.2"
python-dotenv = "^1.0.0"

# Spark processing
pyspark = "^3.4.0"
delta-spark = "^2.4.0"
pyarrow = "^12.0.0"

# Stream processing
confluent-kafka = "^2.1.0"
faust-streaming = "^0.10.16"
# pyflink dependency removed as it's not easily installable via pip

# Database connections
# Downgraded SQLAlchemy to be compatible with Airflow
sqlalchemy = ">=1.4.36,<2.0.0"
psycopg2-binary = "^2.9.5"
google-cloud-bigquery = "^3.9.0"
google-cloud-storage = "^2.8.0"

# Orchestration 
apache-airflow = {version = "^2.6.0", optional = true}
# Kestra client removed as no package exists

# Visualization
streamlit = "^1.21.0"
plotly = "^5.14.1"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"

# Infrastructure
docker = "^6.1.1"
google-api-python-client = "^2.86.0"
google-auth = "^2.17.3"

# DBT
dbt-core = "^1.5.0"
dbt-postgres = "^1.5.0"
dbt-bigquery = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.2.0"
pre-commit = "^3.3.1"
jupyter = "^1.0.0"
notebook = "^6.5.4"
jupyterlab = "^4.0.0"
ipykernel = "^6.22.0"
sphinx = "^7.0.0"
sphinx-rtd-theme = "^1.2.0"
pydocstyle = "^6.3.0"
tox = "^4.5.1"

[tool.poetry.extras]
airflow = ["apache-airflow"]
# kestra extra removed
# flink extra removed
all = ["apache-airflow"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "pyspark.*",
    "confluent_kafka.*",
    "faust.*",
    "dlt.*",
    "google.*",
    "airflow.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=src"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]

[tool.poetry.scripts]
football-pipeline = "src.cli:main"